# specs/001-face-checkin-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FaceCheck API
  version: 1.0.0
  description: |
    Spring Boot 单体后端对 Flutter App 暴露的首版接口契约。
    PostgreSQL 是唯一业务事实源；华为云 FRS 仅作为托管识别能力；
    华为云 OBS 仅作为图片对象存储。
servers:
  - url: https://api.example.com
    description: Production placeholder
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/admin/attendance-records:
    get:
      tags: [AdminSessions]
      summary: List global attendance records
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: sessionId
          schema:
            type: string
            format: uuid
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: status
          schema:
            type: string
            enum: [VALID, MANUAL_CONFIRMED]
      responses:
        '200':
          description: Filtered attendance records returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceRecordSummaryList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/sessions/{sessionId}/records:
    get:
      tags: [AdminSessions]
      summary: List attendance records under a session
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: status
          schema:
            type: string
            enum: [VALID, MANUAL_CONFIRMED]
      responses:
        '200':
          description: Session attendance records returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceRecordSummaryList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/checkin-attempts:
    get:
      tags: [AdminReview]
      summary: List abnormal or filtered check-in attempts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/CheckinStatus'
        - in: query
          name: resultCode
          schema:
            type: string
        - in: query
          name: sessionId
          schema:
            type: string
            format: uuid
        - in: query
          name: reviewed
          schema:
            type: boolean
      responses:
        '200':
          description: Attempt list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCheckinAttemptList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/checkin-attempts/{attemptId}/review:
    post:
      tags: [AdminReview]
      summary: Append review note or mark an abnormal attempt reviewed
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AttemptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCheckinAttemptRequest'
      responses:
        '200':
          description: Review state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCheckinAttemptResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/checkin-attempts/{attemptId}/retry:
    post:
      tags: [AdminReview]
      summary: Retry a failed or manual-review attempt
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AttemptId'
      responses:
        '202':
          description: Retry accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCheckinAttemptResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/system/config:
    get:
      tags: [AdminSystem]
      summary: List editable system configuration
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Config items returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSystemConfigItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/system/config/{configKey}:
    put:
      tags: [AdminSystem]
      summary: Update a system configuration item
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: configKey
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemConfigRequest'
      responses:
        '200':
          description: Config item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSystemConfigItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SessionId:
      in: path
      name: sessionId
      required: true
      schema:
        type: string
        format: uuid
...
    AttemptId:
      in: path
      name: attemptId
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiResponseAttendanceRecordSummaryList:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceRecordSummary'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ApiResponseAdminCheckinAttemptResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/AdminCheckinAttemptResponse'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ApiResponseAdminCheckinAttemptList:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminCheckinAttemptResponse'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    ApiResponseSystemConfigItemList:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/SystemConfigItem'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ApiResponseSystemConfigItem:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/SystemConfigItem'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    CheckinStatus:
      type: string
      enum: [PROCESSING, SUCCESS, FAILED, DUPLICATE_CHECKIN]
...
    AttendanceRecordSummary:
      type: object
      required: [recordId, sessionId, sessionName, userId, checkinTime, status]
      properties:
        recordId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        sessionName:
          type: string
        userId:
          type: string
          format: uuid
        maskedUsername:
          type: string
        checkinTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [VALID, MANUAL_CONFIRMED]
        similarity:
          type: number
          format: double
          nullable: true
    AdminCheckinAttemptResponse:
      type: object
      required: [attemptId, sessionId, sessionName, status, resultCode, createdAt, updatedAt, reviewed, retryCount]
      properties:
        attemptId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        sessionName:
          type: string
        status:
          $ref: '#/components/schemas/CheckinStatus'
        resultCode:
          type: string
        resultMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        matchedUserId:
          type: string
          format: uuid
          nullable: true
        maskedUsername:
          type: string
          nullable: true
        similarity:
          type: number
          format: double
          nullable: true
        reviewed:
          type: boolean
        reviewNote:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reviewedByUserId:
          type: string
          format: uuid
          nullable: true
        retryCount:
          type: integer
...
    SystemConfigItem:
      type: object
      required: [configKey, configValue, valueType]
      properties:
        configKey:
          type: string
        configValue:
          type: string
        valueType:
          type: string
          enum: [STRING, NUMBER, BOOLEAN, JSON]
        description:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
    UpdateSystemConfigRequest:
      type: object
      required: [configValue]
      properties:
        configValue:
          type: string
    ReviewCheckinAttemptRequest:
      type: object
      properties:
        note:
          type: string
          maxLength: 500
        reviewed:
          type: boolean
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
        error:
          $ref: '#/components/schemas/ErrorDetail'
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string